+++ mains/main_make_pdf.py
@@ -1,6 +1,6 @@
 import threading
 
-from src.dlg_config import manage_config
+from src.config import manage_config
 
 
 def create_PDF():
