+++ src/blog_csv_mgr.py
@@ -4,7 +4,7 @@ import os
 from datetime import datetime
 
 from src.aux_res_directory import get_res_folder
-from src.dlg_config import Config, Section, Param
+from src.config import Config, Section, Param
 from src.poster import Poster
