+++ src/config.py
@@ -1,0 +1,190 @@
+import enum
+from configparser import ConfigParser
+from functools import partial
+from pathlib import Path
+from typing import Callable
+
+from src.aux_res_directory import get_res_folder
+from src.dlg_config import DlgConfig
+
+
+SZ_FILE = "General.ini"
+
+
+class Section(enum.Enum):
+    HTML = "HTML"
+    READDATA = "READDATA"
+    COUNT_FILMS = "CONTAR"
+    POST = "POST"
+    DRIVE = "DRIVE"
+
+
+class Param(enum.Enum):
+    # Html
+    FILTER_PUBLISHED = "Filter_published"
+    SCRAP_BLOG = "Force_bog_scraping"
+    ADD_STYLE = "Write_style"
+    OUTPUT_PATH_HTML = "Path_output_html"
+    YES_ALWAYS_DIR = "New_confidence_director"
+    # Readdata
+    FILTER_FA = "Filter_FilmAffinity"
+    DEFAULT_USER = "Mem_user_FA"
+    OUTPUT_EXCEL = "Path_output_excel"
+    # Count films
+    ADD_YEAR = "Add_year"
+    ADD_INDEX = "Add_index"
+    WORD_FOLDER = "Folder_with_words"
+    TITLE_LIST_PATH = "Output_folder_count"
+    # Post
+    BLOG_ID = "Blog_id"
+    DATE = "Posting_date"
+    TIME = "Posting_time"
+    AS_DRAFT = "As_draft"
+    # Drive
+    FOLDER_ID = "Drive_folder_to_update_id"
+    PDF_PATH = "Pdf_folder"
+
+
+def enum_to_str(fun: Callable):
+    '''Convierto los argumentos de enumeración en su valor'''
+
+    def wr(*args):
+
+        # Si el parámetro es Section o Param, lo convierto a string
+        args = tuple(arg.value if type(arg) in (Section, Param)
+                     else arg for arg in args)
+
+        return fun(*args)
+
+    return wr
+
+
+@enum_to_str
+def add_default_value(config: ConfigParser, section: Section, param: Param, value) -> None:
+
+    # Si no existe la sección, la añado
+    if section not in config:
+        config.add_section(section)
+    # Si no existe el parámetro lo añado con el valor default
+    if param not in config[section]:
+        config.set(section, param, str(value))
+
+
+class Config:
+
+    # Abro el lector del archivo
+    config = ConfigParser()
+    # Dirección del ini
+    sz_path = get_res_folder(SZ_FILE)
+    config.read(sz_path, encoding="utf-8")
+
+    dlg_config = DlgConfig(config)
+
+    add_def_value = partial(add_default_value, config)
+    # Configuraciones para html
+    add_def_value(Section.HTML, Param.FILTER_PUBLISHED, False)
+    add_def_value(Section.HTML, Param.SCRAP_BLOG, False)
+    add_def_value(Section.HTML, Param.ADD_STYLE, False)
+    add_def_value(Section.HTML, Param.OUTPUT_PATH_HTML, 'auto')
+    add_def_value(Section.HTML, Param.YES_ALWAYS_DIR, "")
+    # Configuraciones para readdata
+    add_def_value(Section.READDATA, Param.FILTER_FA, 1)
+    add_def_value(Section.READDATA, Param.DEFAULT_USER, 'Jorge')
+    add_def_value(Section.READDATA, Param.OUTPUT_EXCEL, 'auto')
+    # Configuraciones para escribir el txt
+    add_def_value(Section.COUNT_FILMS, Param.ADD_YEAR, False)
+    add_def_value(Section.COUNT_FILMS, Param.ADD_INDEX, False)
+    add_def_value(Section.COUNT_FILMS, Param.WORD_FOLDER, 'auto')
+    add_def_value(Section.COUNT_FILMS,
+                  Param.TITLE_LIST_PATH, 'auto')
+    # Configuraciones para post
+    add_def_value(Section.POST, Param.BLOG_ID,
+                  '4259058779347983900')
+    add_def_value(Section.POST, Param.DATE, 'auto')
+    add_def_value(Section.POST, Param.TIME, '20:00')
+    add_def_value(Section.POST, Param.AS_DRAFT, False)
+    # Configuraciones para actualizar drive
+    add_def_value(Section.DRIVE, Param.FOLDER_ID,
+                  '13UbwzbjVFQ8e_UaNalqm_iMihjBDBvtm')
+    add_def_value(Section.DRIVE, Param.PDF_PATH, 'auto')
+
+    @classmethod
+    def save_config(cls):
+        with open(cls.sz_path, 'w', encoding="utf-8") as configfile:
+            cls.config.write(configfile)
+
+    @classmethod
+    @enum_to_str
+    def get_value(cls, section: Section, param: Param) -> str:
+        return cls.config[section][param]
+
+    @classmethod
+    @enum_to_str
+    def get_int(cls, section: Section, param: Param) -> int:
+        return cls.config.getint(section, param)
+
+    @classmethod
+    @enum_to_str
+    def get_bool(cls, section: Section, param: Param) -> bool:
+        return cls.config.getboolean(section, param)
+
+    @classmethod
+    @enum_to_str
+    def get_folder_path(cls, section: Section, param: Param) -> Path:
+        # Leo lo que hya escrito en el ini
+        ini_data = cls.config[section][param]
+
+        # Compruebo que sea una carpeta
+        while not Path(ini_data).is_dir():
+            # Si no es una carpeta válida, la pido al usuario
+            ini_data = input(
+                f"Introducir path de la carpeta {section} {param}: ")
+            # Guardo el dato elegido
+            cls.set_value(section, param, ini_data)
+
+        return Path(ini_data)
+
+    @classmethod
+    @enum_to_str
+    def get_file_path(cls, section: Section, param: Param) -> Path:
+        # Leo lo que hya escrito en el ini
+        ini_data = cls.config[section][param]
+
+        # Compruebo que sea una carpeta
+        while not Path(ini_data).is_file():
+            # Si no es una carpeta válida, la pido al usuario
+            ini_data = input(
+                f"Introducir path del archivo {section} {param}: ")
+            # Guardo el dato elegido
+            cls.set_value(section, param, ini_data)
+
+        return Path(ini_data)
+
+    @classmethod
+    @enum_to_str
+    def set_value(cls, section: Section, param: Param, value) -> None:
+        # Me espero que se introduzca un valor en una sección que existe
+        if param not in cls.config[section]:
+            assert("{} no pertenece a la sección {}.".format(param, section))
+
+        # Lo cambio en el objeto
+        cls.config.set(section, param, str(value))
+
+        # Actualizo el archivo ini
+        cls.save_config()
+
+    @classmethod
+    def run_dlg(cls):
+        cls.dlg_config.run()
+
+
+def manage_config():
+    # Importo los módulos de windows para comprobar el teclado
+    import win32api
+    import win32con
+
+    # Comrpuebo si la tecla control está apretada
+    if win32api.GetAsyncKeyState(win32con.VK_CONTROL) & 0x8000 > 0:
+        # Abro el diálogo
+        Config.run_dlg()
+        Config.save_config()
