+++ src/content_mgr.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from bs4 import BeautifulSoup
 
 from src.aux_title_str import split_title_year
-from src.dlg_config import Config, Section, Param
+from src.config import Config, Section, Param
 from src.dlg_scroll_base import DlgScrollBase
 from src.make_html import SZ_HTML_FILE
 from src.read_blog import BlogHiddenData
