+++ src/pdf_writer.py
@@ -3,7 +3,7 @@ import os
 import docx2pdf
 from PyPDF2 import PdfFileMerger
 
-from src.dlg_config import Config, Section, Param
+from src.config import Config, Section, Param
 from src.word_folder_mgr import WordFolderMgr
